feat: stack `title` and `description` without overlap when sharing position
- Introduced `BlockLayout` class for measuring and rendering stacked text blocks.
- Updated `Generator` to group and align blocks sharing the same position as a single unit.
- Added vertical spacing logic to avoid overlapping stacked blocks.
- Enhanced tests and added new scenarios for validating stacked block rendering.
- Updated snapshots to reflect layout changes.

+++ src/Text/Writer.php
@@ -65,8 +65,8 @@ final class Writer
             throw new RuntimeException('Failed to calculate text bounding box');
         }
 
-        $ascent = -$bbox[7];  // top of glyph above baseline (bbox[7] is negative)
-        $descent = $bbox[1];  // below baseline
+        $ascent = (int) -$bbox[7];  // top of glyph above baseline (bbox[7] is negative)
+        $descent = (int) $bbox[1];  // below baseline
 
         return ['ascent' => $ascent, 'height' => $ascent + $descent];
     }
